[HttpFoundation] moved test file to the right directory
